Tentar excluir só se existir...

+++ transcreve_podcast_feed.py
@@ -79,6 +79,7 @@ for i in feed.entries:
         arq_saida = open('{}/{}'.format(sys.argv[2].replace('/', '_'), nome_arq_saida), 'w')
         arq_saida.write(srt.compose(transcribe()))
         arq_saida.close()
-    os.remove('/tmp/arq_media_podcast.{}'.format(sys.argv[2].replace('/', '_')))
+    if os.path.exists('/tmp/arq_media_podcast.{}'.format(sys.argv[2].replace('/', '_'))):
+        os.remove('/tmp/arq_media_podcast.{}'.format(sys.argv[2].replace('/', '_')))
 
 os.remove('/tmp/semaforo_transcreve_podcast_feed.{}'.format(sys.argv[2].replace('/', '_')))
